cis azure 3.11
